feat: Add model for `RetrieveHolidays`
`Request.RetrieveHolidays` is a struct that I use to encode an action
in. One can say it's not really necessary, since I could just use a map,
and to a certain extent, that is true. But I want to be able to narrow
down things even further so that it removes any wiggle room for
mistakes.

So, ideally, any function that needs to handle the `RetrieveHolidays`
request should just set it as its input in the type specs, or even
pattern match against it, and you'll be given the guarantee that
all fields are present.

Not exactly as good as languages with algebraic data types, but good
enough, I suppose.