feat: Add `AddCustomHoliday`
This represents the request to add a custom holiday. It does basic
validations like checking if the year is not negative, or if the
range follows a certain property. But it currently does not validate if
the day of a certain month is valid.

Note: This is still a draft.